sync

+++ INF/B7/Agiles Softwareentwicklungsprojekt (ASEP)/AssignmentExample1/Vector2D.cs
@@ -2,13 +2,13 @@
 
 internal class Vector2D
 {
-    private int x;
-    public int X { get => x; set => x = value; }
+    private double x;
+    public double X { get => x; set => x = value; }
 
-    private int y;
-    public int Y { get => y; set => y = value; }
+    private double y;
+    public double Y { get => y; set => y = value; }
 
-    public Vector2D(int x = 0, int y = 0)
+    public Vector2D(double x = 0, double y = 0)
     {
         this.X = x;
         this.Y = y;
@@ -26,8 +26,8 @@ internal class Vector2D
 
     public static double Distance(Vector2D v1, Vector2D v2)
     {
-        int x =  v2.X- v1.X;
-        int y = v2.Y - v1.Y;
+        double x =  v2.X- v1.X;
+        double y = v2.Y - v1.Y;
 
         double result = Math.Sqrt(Math.Pow(x, 2) + Math.Pow(y, 2));
